+++ packages/bignumber/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "bignumber/5.0.10";
+export declare const version = "bignumber/5.0.11";