+++ packages/cli/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "cli/5.0.6";
+export declare const version = "cli/5.0.7";